updated symmetric joukowski
Tried to rectify the pressure plot by changing the number of panels at the trailing edge.  As you remove panels at the trailing edge the  pressure plots approach similar plots in Katz and Plotkin but the CL decreases do to decreasing chord length.  I tried to  lengthen the chord then take off panels at the end but CL still is decreased do to the TE not being a cusp.